requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
dotenv
request